+++ modules/codegen/src/smithy4s/codegen/SmithyBuildJson.scala
@@ -21,11 +21,13 @@ import smithy4s.codegen.internals.SmithyBuildMaven
 import smithy4s.codegen.internals.SmithyBuildMavenRepository
 import io.circe.{Json, parser}
 
+import scala.collection.immutable.ListSet
+
 private[codegen] object SmithyBuildJson {
   def toJson(
-      imports: Seq[String],
-      dependencies: Seq[String],
-      repositories: Seq[String]
+      imports: ListSet[String],
+      dependencies: ListSet[String],
+      repositories: ListSet[String]
   ): String = {
     SmithyBuild.writeJson(
       SmithyBuild(